+++ contrib/opencensus-ext-azure/opencensus/ext/azure/common/__init__.py
@@ -92,6 +92,7 @@ class Options(BaseObject):
         export_interval=15.0,
         grace_period=5.0,
         instrumentation_key=None,
+        logging_sampling_rate=1.0,
         max_batch_size=100,
         minimum_retry_interval=60,  # minimum retry interval in seconds
         proxy=None,