+++ contrib/opencensus-ext-requests/tests/test_requests_trace.py
@@ -16,11 +16,12 @@ import unittest
 
 import mock
 import requests
-from opencensus.trace.tracers import noop_tracer
 
 from opencensus.ext.requests import trace
-from opencensus.trace import span as span_module, execution_context
+from opencensus.trace import execution_context
+from opencensus.trace import span as span_module
 from opencensus.trace import status as status_module
+from opencensus.trace.tracers import noop_tracer
 
 
 class Test_requests_trace(unittest.TestCase):
@@ -103,6 +104,7 @@ class Test_requests_trace(unittest.TestCase):
             wrapped(url)
 
         expected_attributes = {
+            'component': 'HTTP',
             'http.host': 'localhost:8080',
             'http.method': 'GET',
             'http.path': '/test',
@@ -246,6 +248,7 @@ class Test_requests_trace(unittest.TestCase):
                 wrapped(url)
 
         expected_attributes = {
+            'component': 'HTTP',
             'http.host': 'localhost:8080',
             'http.method': 'GET',
             'http.path': '/test',
@@ -293,6 +296,7 @@ class Test_requests_trace(unittest.TestCase):
                 wrapped(url)
 
         expected_attributes = {
+            'component': 'HTTP',
             'http.host': 'localhost:8080',
             'http.method': 'GET',
             'http.path': '/test',
@@ -341,6 +345,7 @@ class Test_requests_trace(unittest.TestCase):
                 wrapped(url)
 
         expected_attributes = {
+            'component': 'HTTP',
             'http.host': 'localhost:8080',
             'http.method': 'GET',
             'http.path': '/test',
@@ -387,6 +392,7 @@ class Test_requests_trace(unittest.TestCase):
             )
 
         expected_attributes = {
+            'component': 'HTTP',
             'http.host': 'localhost:8080',
             'http.method': 'POST',
             'http.path': '/test',
@@ -618,6 +624,7 @@ class Test_requests_trace(unittest.TestCase):
                 )
 
         expected_attributes = {
+            'component': 'HTTP',
             'http.host': 'localhost:8080',
             'http.method': 'POST',
             'http.path': '/test',
@@ -666,6 +673,7 @@ class Test_requests_trace(unittest.TestCase):
                 )
 
         expected_attributes = {
+            'component': 'HTTP',
             'http.host': 'localhost:8080',
             'http.method': 'POST',
             'http.path': '/test',
@@ -714,6 +722,7 @@ class Test_requests_trace(unittest.TestCase):
                 )
 
         expected_attributes = {
+            'component': 'HTTP',
             'http.host': 'localhost:8080',
             'http.method': 'POST',
             'http.path': '/test',